feat: products seeder

+++ database/migrations/2024_06_03_154217_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 150);
             $table->string('description', 500);
-            $table->tinyInteger('price');
+            $table->float('price');
             $table->string('image');
             $table->timestamps();
         });
